Refactor archivos cobrar presentacion
